pyproject.toml
[tool.poetry]
name = "realestatews"
version = "0.1.0"
description = ""
authors = ["Leonid Sliusar <loslyusar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
motor = "^3.3.1"
pydantic-settings = "^2.0.3"
pydantic = "^2.3.0"
uvicorn = "^0.23.2"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
pdf2image = "^1.16.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = "app"
testpaths = [
    "tests",
]